fix: BrandStory 디렉토리 생성 및 파일 분리 (#67) - BrandStory 디렉토리 생성 - HomeStyle => BrandStoryStyle로 파일 분리
